HW-1 удалил все файлы из проекта которые должен написать самостоятельно
