+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/api/SpendApiClient.java
@@ -1,124 +1,33 @@
 package guru.qa.niffler.api;
 
 import guru.qa.niffler.config.Config;
-import guru.qa.niffler.model.CategoryJson;
-import guru.qa.niffler.model.CurrencyValues;
 import guru.qa.niffler.model.SpendJson;
 import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.jackson.JacksonConverterFactory;
 
-import javax.annotation.Nonnull;
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class SpendApiClient {
 
-  private final Retrofit retrofit = new Retrofit.Builder()
-      .baseUrl(Config.getInstance().spendUrl())
-      .addConverterFactory(JacksonConverterFactory.create())
-      .build();
+    private final Retrofit retrofit = new Retrofit.Builder()
+            .baseUrl(Config.getInstance().spendUrl())
+            .addConverterFactory(JacksonConverterFactory.create())
+            .build();
 
-  private final SpendApi spendApi = retrofit.create(SpendApi.class);
+    private final SpendApi spendApi = retrofit.create(SpendApi.class);
 
-  public SpendJson createSpend(SpendJson spend) {
-    final Response<SpendJson> response;
-    try {
-      response = spendApi.addSpend(spend)
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
+    public SpendJson createSpend(SpendJson spend) {
+        final Response<SpendJson> response;
+        try {
+            response = spendApi.addSpend(spend)
+                    .execute();
+        } catch (IOException e) {
+            throw new AssertionError(e);
+        }
+        assertEquals(201, response.code());
+        return response.body();
     }
-    assertEquals(201, response.code());
-    return response.body();
-  }
-
-  public SpendJson editSpend(SpendJson spend) {
-    final Response<SpendJson> response;
-    try {
-      response = spendApi.editSpend(spend)
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
-    }
-    assertEquals(200, response.code());
-    return response.body();
-  }
-
-  public SpendJson getSpend(String id) {
-    final Response<SpendJson> response;
-    try {
-      response = spendApi.getSpend(id)
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
-    }
-    assertEquals(200, response.code());
-    return response.body();
-  }
-
-  public List<SpendJson> allSpends(String username,
-                                   CurrencyValues currency,
-                                   String from,
-                                   String to) {
-    final Response<List<SpendJson>> response;
-    try {
-      response = spendApi.allSpends(username, currency, from, to)
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
-    }
-    assertEquals(200, response.code());
-    return response.body();
-  }
-
-  public void removeSpends(@Nonnull String username, @Nonnull String... ids) {
-    final Response<Void> response;
-    try {
-      response = spendApi.removeSpends(username, Arrays.stream(ids).toList())
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
-    }
-    assertEquals(200, response.code());
-  }
-
-  public CategoryJson createCategory(CategoryJson category) {
-    final Response<CategoryJson> response;
-    try {
-      response = spendApi.addCategory(category)
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
-    }
-    assertEquals(200, response.code());
-    return response.body();
-  }
-
-  public CategoryJson updateCategory(CategoryJson category) {
-    final Response<CategoryJson> response;
-    try {
-      response = spendApi.updateCategory(category)
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
-    }
-    assertEquals(200, response.code());
-    return response.body();
-  }
-
-  public List<CategoryJson> allCategory(String username) {
-    final Response<List<CategoryJson>> response;
-    try {
-      response = spendApi.allCategories(username)
-          .execute();
-    } catch (IOException e) {
-      throw new AssertionError(e);
-    }
-    assertEquals(200, response.code());
-    return response.body();
-  }
 }
