+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/config/LocalConfig.java
@@ -5,7 +5,7 @@ enum LocalConfig implements Config {
 
   @Override
   public String frontUrl() {
-    return "http://127.0.0.1:3000/";
+    return "http://127.0.0.1:9000/";
   }
 
   @Override
