--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/Category.java
@@ -1,19 +1,0 @@
-package guru.qa.niffler.jupiter.annotation;
-
-import guru.qa.niffler.jupiter.extension.CategoryExtension;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-@ExtendWith(CategoryExtension.class)
-public @interface Category {
-  String username();
-
-  boolean archived() default false;
-
-}
