+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/meta/WebTest.java
@@ -1,7 +1,6 @@
 package guru.qa.niffler.jupiter.annotation.meta;
 
 import guru.qa.niffler.jupiter.extension.BrowserExtension;
-import guru.qa.niffler.jupiter.extension.UsersQueueExtension;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import java.lang.annotation.ElementType;
@@ -12,8 +11,7 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 @ExtendWith({
-    BrowserExtension.class,
-    UsersQueueExtension.class
+        BrowserExtension.class,
 })
 public @interface WebTest {
 }
