+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/SpendingResolverExtension.java
@@ -1,0 +1,19 @@
+package guru.qa.niffler.jupiter.extension;
+
+import guru.qa.niffler.model.SpendJson;
+import org.junit.jupiter.api.extension.*;
+
+public class SpendingResolverExtension implements ParameterResolver {
+
+    public static final ExtensionContext.Namespace NAMESPACE = ExtensionContext.Namespace.create(SpendingResolverExtension.class);
+
+    @Override
+    public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return parameterContext.getParameter().getType().isAssignableFrom(SpendJson.class);
+    }
+
+    @Override
+    public SpendJson resolveParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return extensionContext.getStore(CreateSpendingExtension.NAMESPACE).get(extensionContext.getUniqueId(), SpendJson.class);
+    }
+}
