--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/page/FriendsPage.java
@@ -1,30 +1,0 @@
-package guru.qa.niffler.page;
-
-import com.codeborne.selenide.SelenideElement;
-
-import static com.codeborne.selenide.CollectionCondition.size;
-import static com.codeborne.selenide.CollectionCondition.textsInAnyOrder;
-import static com.codeborne.selenide.Selenide.$;
-
-public class FriendsPage {
-
-  private final SelenideElement peopleTab = $("a[href='/people/friends']");
-  private final SelenideElement allTab = $("a[href='/people/all']");
-  private final SelenideElement requestsTable = $("#requests");
-  private final SelenideElement friendsTable = $("#friends");
-
-  public FriendsPage checkExistingFriends(String... expectedUsernames) {
-    friendsTable.$$("tr").shouldHave(textsInAnyOrder(expectedUsernames));
-    return this;
-  }
-
-  public FriendsPage checkNoExistingFriends() {
-    friendsTable.$$("tr").shouldHave(size(0));
-    return this;
-  }
-
-  public FriendsPage checkExistingInvitations(String... expectedUsernames) {
-    requestsTable.$$("tr").shouldHave(textsInAnyOrder(expectedUsernames));
-    return this;
-  }
-}
