+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/page/MainPage.java
@@ -1,7 +1,6 @@
 package guru.qa.niffler.page;
 
 import com.codeborne.selenide.ElementsCollection;
-import com.codeborne.selenide.SelenideElement;
 
 import static com.codeborne.selenide.Condition.text;
 import static com.codeborne.selenide.Condition.visible;
@@ -9,23 +8,7 @@ import static com.codeborne.selenide.Selenide.$;
 
 public class MainPage {
 
-  private final SelenideElement header = $("#root header");
-  private final SelenideElement headerMenu = $("ul[role='menu']");
   private final ElementsCollection tableRows = $("#spendings tbody").$$("tr");
-  private final SelenideElement statComponent = $("#stat");
-  private final SelenideElement spendingTable = $("#spendings");
-
-  public FriendsPage friendsPage() {
-    header.$("button").click();
-    headerMenu.$$("li").find(text("Friends")).click();
-    return new FriendsPage();
-  }
-
-  public PeoplePage allPeoplesPage() {
-    header.$("button").click();
-    headerMenu.$$("li").find(text("All People")).click();
-    return new PeoplePage();
-  }
 
   public EditSpendingPage editSpending(String spendingDescription) {
     tableRows.find(text(spendingDescription)).$$("td").get(5).click();
@@ -34,11 +17,5 @@ public class MainPage {
 
   public void checkThatTableContainsSpending(String spendingDescription) {
     tableRows.find(text(spendingDescription)).should(visible);
-  }
-
-  public MainPage checkThatPageLoaded() {
-    statComponent.should(visible).shouldHave(text("Statistics"));
-    spendingTable.should(visible).shouldHave(text("History of Spendings"));
-    return this;
   }
 }
