--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/page/PeoplePage.java
@@ -1,19 +1,0 @@
-package guru.qa.niffler.page;
-
-import com.codeborne.selenide.SelenideElement;
-
-import static com.codeborne.selenide.Condition.text;
-import static com.codeborne.selenide.Selenide.$;
-
-public class PeoplePage {
-
-  private final SelenideElement peopleTab = $("a[href='/people/friends']");
-  private final SelenideElement allTab = $("a[href='/people/all']");
-  private final SelenideElement peopleTable = $("#all");
-
-  public PeoplePage checkInvitationSentToUser(String username) {
-    SelenideElement friendRow = peopleTable.$$("tr").find(text(username));
-    friendRow.shouldHave(text("Waiting..."));
-    return this;
-  }
-}
