--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/LoginTest.java
@@ -1,29 +1,0 @@
-package guru.qa.niffler.test.web;
-
-import com.codeborne.selenide.Selenide;
-import com.github.javafaker.Faker;
-import guru.qa.niffler.config.Config;
-import guru.qa.niffler.jupiter.annotation.meta.WebTest;
-import guru.qa.niffler.page.LoginPage;
-import org.junit.jupiter.api.Test;
-
-@WebTest
-public class LoginTest {
-
-  private static final Config CFG = Config.getInstance();
-  private static final Faker faker = new Faker();
-
-  @Test
-  void mainPageShouldBeDisplayedAfterSuccessLogin() {
-    Selenide.open(CFG.frontUrl(), LoginPage.class)
-        .successLogin("duck", "12345")
-        .checkThatPageLoaded();
-  }
-
-  @Test
-  void userShouldStayOnLoginPageAfterLoginWithBadCredentials() {
-    LoginPage loginPage = Selenide.open(CFG.frontUrl(), LoginPage.class);
-    loginPage.login(faker.name().username(), "BAD");
-    loginPage.checkError("Bad credentials");
-  }
-}
