--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/ProfileTest.java
@@ -1,44 +1,0 @@
-package guru.qa.niffler.test.web;
-
-import com.codeborne.selenide.Selenide;
-import guru.qa.niffler.config.Config;
-import guru.qa.niffler.jupiter.annotation.Category;
-import guru.qa.niffler.jupiter.annotation.meta.WebTest;
-import guru.qa.niffler.model.CategoryJson;
-import guru.qa.niffler.page.LoginPage;
-import guru.qa.niffler.page.ProfilePage;
-import org.junit.jupiter.api.Test;
-
-@WebTest
-public class ProfileTest {
-
-  private static final Config CFG = Config.getInstance();
-
-  @Category(
-      username = "duck",
-      archived = true
-  )
-  @Test
-  void archivedCategoryShouldPresentInCategoriesList(CategoryJson category) {
-    Selenide.open(CFG.frontUrl(), LoginPage.class)
-        .successLogin("duck", "12345")
-        .checkThatPageLoaded();
-
-    Selenide.open(CFG.frontUrl() + "profile", ProfilePage.class)
-        .checkArchivedCategoryExists(category.name());
-  }
-
-  @Category(
-      username = "duck",
-      archived = false
-  )
-  @Test
-  void activeCategoryShouldPresentInCategoriesList(CategoryJson category) {
-    Selenide.open(CFG.frontUrl(), LoginPage.class)
-        .successLogin("duck", "12345")
-        .checkThatPageLoaded();
-
-    Selenide.open(CFG.frontUrl() + "profile", ProfilePage.class)
-        .checkCategoryExists(category.name());
-  }
-}
