+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/test/web/SpendingWebTest.java
@@ -2,18 +2,20 @@ package guru.qa.niffler.test.web;
 
 import com.codeborne.selenide.Selenide;
 import guru.qa.niffler.config.Config;
-import guru.qa.niffler.jupiter.extension.BrowserExtension;
+import guru.qa.niffler.jupiter.annotation.meta.WebTest;
 import guru.qa.niffler.jupiter.annotation.Spending;
+import guru.qa.niffler.jupiter.extension.BrowserExtension;
 import guru.qa.niffler.model.SpendJson;
 import guru.qa.niffler.page.LoginPage;
 import guru.qa.niffler.page.MainPage;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
+@WebTest
 @ExtendWith(BrowserExtension.class)
 public class SpendingWebTest {
 
-  private static final Config CFG = Config.getInstanceDocker();
+  private static final Config CFG = Config.getInstance();
 
   @Spending(
       username = "duck",
@@ -26,7 +28,7 @@ public class SpendingWebTest {
     final String newDescription = "Обучение Niffler Next Generation";
 
     Selenide.open(CFG.frontUrl(), LoginPage.class)
-        .login("cat", "12345")
+        .successLogin("duck", "12345")
         .editSpending(spend.description())
         .setNewSpendingDescription(newDescription)
         .save();
