--- niffler-e-2-e-tests/src/test/resources/META-INF/services/org.junit.jupiter.api.extension.Extension
@@ -1,1 +1,0 @@
-guru.qa.niffler.jupiter.extension.TestMethodContextExtension
